Address code review: extract JWT util, validate product ID, add product_manager access tests
