+++ frontend/src/pages/sales-manager/SalesManagerDashboard.jsx
@@ -1,18 +1,7 @@
 import { useState } from 'react'
 import DashboardLayout from '../../components/DashboardLayout'
 import PriceManagement from './PriceManagement'
-
-function decodeJwtPayload(token) {
-  try {
-    const parts = token.split('.')
-    if (parts.length < 3) return null
-    const base64 = parts[1].replace(/-/g, '+').replace(/_/g, '/')
-    const padded = base64 + '='.repeat((4 - (base64.length % 4)) % 4)
-    return JSON.parse(atob(padded))
-  } catch {
-    return null
-  }
-}
+import { decodeJwtPayload } from '../../utils/jwt'
 
 const sections = [{ key: 'products', label: 'Products', icon: <PriceTagIcon /> }]
