+++ frontend/src/pages/sales-manager/SalesManagerLoginPage.jsx
@@ -3,18 +3,7 @@ import API_BASE from '../../api'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
-
-function decodeJwtPayload(token) {
-  try {
-    const parts = token.split('.')
-    if (parts.length < 3) return null
-    const base64 = parts[1].replace(/-/g, '+').replace(/_/g, '/')
-    const padded = base64 + '='.repeat((4 - (base64.length % 4)) % 4)
-    return JSON.parse(atob(padded))
-  } catch {
-    return null
-  }
-}
+import { decodeJwtPayload } from '../../utils/jwt'
 
 function SalesManagerLoginPage({ onLogin }) {
   const [email, setEmail] = useState('')
