+++ frontend/src/utils/jwt.js
@@ -1,0 +1,11 @@
+export function decodeJwtPayload(token) {
+  try {
+    const parts = token.split('.')
+    if (parts.length < 3) return null
+    const base64 = parts[1].replace(/-/g, '+').replace(/_/g, '/')
+    const padded = base64 + '='.repeat((4 - (base64.length % 4)) % 4)
+    return JSON.parse(atob(padded))
+  } catch {
+    return null
+  }
+}
